Pass memory to clist_make_accum_device()
Change function args to propagate the memory type of the graphics
state to clist_make_accum_device to create device with it.

Also, make sure the band_memory is allocated in the mem->non_gc_memory
instead of the parent device's (fixes crash in gpdl/pcl that was
introduced).

That will mean pdfi and gs/pdfi get non-gc, and gs gets gc.

This reverts previous commit (on pdfi branch) that was forcing it to
use non-gc in all cases.  I shouldn't have pushed that commit because
it turned out it caused issues with plain gs.

There was a bug in gs/pdfi where the clist-pattern device was in gc
memory and it was getting freed when pdfi was still using it.

This should fix a segfault in: tests_private/comparefiles/Bug688396.pdf