Fix refcnt's on the colorspace in clist
In the case where the colorspace was modified (read_set_color_space)
there was a memory leak.

This was because the old code just freed pcs, but if the color had
changed, then gs_state.color[0].color_space and [1].color_space are no
longer the same.  This way we track the references to them separately
and free them separately.

pcs doesn't get its own ref, because it is always the same as
gs_gstate.color[0].color_space.

This fixes memory leaks that only show up in the case that the memory
is not garbage-collected, such as in pdfi.