Improve eps2write
Bug #702521 "eps2write hygiene enhancements"

If processing a multi-page input file, throw an error on the second
page and write a message explaining the problem.

Encapsulate the 'global configuration' switches which affect the prolog
processing, and are written from the ps2write device, in a dictionary
because EPS files should not affect the current dictionary. Read those
parameters back from the dictionary when establishing the EPS file's
own dictionary, ensuring that a dictionary is actually present.

Use the global configuration to determine whether to install a custom
error handler. If its an EPS file, then don't.