Slight tweak to the FAPI API
The previous Font API update had a slight mistake: for non-CID fonts, we'd pass
GS_NO_GLYPH in place of the actually CID/GID into the "fapi_set_cache" call
back.

That worked fine because, for Postscript (and currrent PDF) non-CID fonts use
the glyph name, not a character code, and the other interpreters do not
need a character for that callback.

We'll now pass the glyph index for non-CID fonts, and the
glyph index + GS_MIN_CID_GLYPH (as is conventional for gs) - and that allows
the fapi_set_cache call back to differentiate between CID and non-CID fonts.

For the new PDF interpreter (pdfi) we actually do need the character code, to
correctly handle Truetype fonts. So fixing the graphics lib and associated code
on master, so the pdfi branch can also work.