Fix heap corruption with Revision 5 or 6 encryption and short passwords.
We allocate a buffer to hold a string partially built from the password
which we then use as an MD5 buffer. The problem was that if the password
was too short (6 or fewer characters) then the buffer we allocated
wouldn't be large enough to hold the 128 bits of the MD5, we would
write off the end of the buffer causing heap corruption.

Make sure the buffer is at least 16 bytes long.