Fix parsing of number to allow for an exponent
Changed logic to allow an 'e' or an 'E' with a +/- sign.
Also changed the brackets/indentation of the if/else to be more readable.

Added a warning message for this, since it isn't standard PDF spec.
gs and other readers (including mupdf, evince) do accept the
scientific notation.

This fixes: tests_private/pdf/sumatra/1643_-_ignore_broken_arrays_inside_dicts.pdf

This file has: "/CropBox [0 0 5e-02 0]"

Even when the file renders without errors, since the CropBox is tiny,
many viewers render nothing.  mupdf and evince seem to ignore the CropBox.

It seems pdfi is currently not dealing with the CropBox the same way as gs
does.  Should revisit that.