pdfwrite - when copying clip path, cater for differing allocators
This is for pdfi when run under Ghostscript; the pdfwrite device keeps
a copy of the current clip path when it writes it to the output. But
the code (gx_cpath_to_path) relies on gx_path_asign_preserve and that
function specifically states that the allocators for the two paths
must be the same.

In the case of pdfi under GS, the path is created using the pdfi memory
allocator (its created by the pdfi interpreter), but the copied path
is allocated using the device allocator, which is derived from the
PostScript interpreter. The result of this is that pdfwrite can still
be pointing to a path which was allocated by pdfi and can disappear
when the pdfi interpreter exits, leaving a dangling pointer. Since the
path is subject to garbage collection (allocated with the PostScript
interpreter allocator) this causes a crash as soon as any kind of
grbage collection takes place.

Fortunately there's a simple function call 'gx_path_unshare' which will
copy the segments after assignment and we can check the allocators to
see if they are the same and us this function if they are not.

Fixes the remaining (currently) seg faults with pdf when running inside
Ghostscript.