Improve PostScript integration
The newpdf_set_page routine was handling the case of the MediaBox not
having its origin at 0,0 by translating the CTM, but it was doing so
before setting the PageSize, and the call to setpaegdevice reset the
CTM, so the translation was lost. It also wasn't catering for the case
where the MediaBox was negative (IA3Z0005.pdf).

Added a routine to gather interpreter parameters relevant to the PDF
interpreter, create a dictionary from them in order to pass to the
interpreter.

More reworking of the interface routines; added a new .PDFInit which
takes a dictionary as a parameter and uses its contents to configure
the PDF interpreter. This routine now creates the PDF interpreter
context and returns the pdffile object (with no file pointers).

Removed the context creation from .PDFFile and .PDFStream, these now
require the -pdffile- containing the context to be passed as a parameter

This reduces the number of errors when running pdfi from Ghostscript
to ~800 (from > 3,500). Almost all the remaining problems are with the
ps2write device and font related.