Rearrange the graphic state setup code
This is trying to straighten out the current messy situation.

For pdfwrite device, the order of these things can actually matter
because it changes what is emitted in the pdf file output.

Some things need to be set up in pdf_impl_set_device() before the
pdfwrite device is initialized, otherwise it will generate extra
entries in various ExtGState dictionaries.  These are because of a
mismatch in what the device thinks the current gstate is, vs. what the
gstate actually is.

It's mostly harmless, but for example setting "/SM .02" seems wrong --
it should not set it at all in most circumstances, and just be using
the default (which is determined by the pdf file-consumer -- gs
happens to use .02, but our pdfwrite output shouldn't enforce that).

Made a function pdfi_page_graphics_begin() which collects a bunch of
the setup that needs to be done in both pdfi and gs/pdfi code, as well
as for annotations.

This whole graphics/page setup situation is still a bit of a mess, but
hopefully this is progress.

I am testing this by comparing pdfwrite output between gs, gpdf and
gs/pdfi and looking for oddities in the ExtGState's.