Bug 702553: Broken handling of 1.0 TTF cmap case
A (clearly!) fairly little used path through the TTF cmap table handling code
for PDF files had bitrotted to the point where we ended up trying to create
a Type 42 font (from a TTF) which didn't have a CharStrings dictionary.

Specifically, this is the case where we're using a 1.0 cmap table, and the
source PDF font's Encoding object does not contain a Differences key.