pdfwrite - move TwoByteToUnicode variable out of union
Discovered while working on the new PDF interpreter and the test file
BCL_Easy.pdf with pdfwrite.

The TwoByteToUnicode flag was stored in the 'simple' structure in the
encoding union. However, that flag is altered by the addition of
ToUnicode code points, and that is *not* limited to simple encodings,
we also alter it for CIDFonts. If we alter the flag we are, in effect,
corrupting members of the union in the case where we are using a
different structure in the union.

This showed up as corrupting the CMapName member of the type0 structure
when built for 64-bit. It didn't exhibit with regular Ghostscript
because that uses the ToUnicode CMap from the input, and only has a
2-byte Unicode value, the pdfi version currently returns a 4-byte
code point.

Nevertheless, this clearly shows that the potential for problems is
there, so to avoid this in future move the flag out of the union and
into the main pdf_font_resource_s structure.