Add a 'get_glyph_index' function to TrueType fonts
The pdfwrite device wants to turn a character code into a GID when
it marks a glyph as being used by the input file. Previously we were
leaving this with the default implementation which doesn't do anything
useful.

I've chosen to add a function to set the procedure in the gs_font_type42
structure, because I think its possible we'll need to modify some other
function pointers in the future.

The actual work is done by pdfi_type42_get_glyph_index() which calls the
existing pdfi_fapi_check_cmap_for_GID().

This causes many pages of diffs with pdfwrite. There are some files
which still display the wrong text (or incorrect spacing) but that's
almost certainly a different problem.