pdfwrite - write FontMatrix for FDArray of a CFF CIDFont
Bug #702556 "pdfwrite produces an invalid pdf file"

The description is incorrect here, the PDF file is not invalid. The
input PostScript file has been created by converting a PDF file using
Poppler.

The original PDF file contains a number of CIDFonts, all with CFF
outlines, and the conversion, for reasons best known to itself, moves
the FontMatrix out of the CFF font and into the CIDFont. This isn't
possible with PDF, so pdfwrite is forced to move the FontMatrix from
the CIDFont back into the CFF font.

Unfortunately we were not writing the FontMatrix out at the FDArray
level of the CFF CIDFont, leading to text 1000 times too large.

Fixed here by writing the FontMatrix for each font in the FDArray.