Bug 702560: Cope with (broken) form with "Contents"
The file for this bug has a form which contains a /Contents key for the content
stream, rather than the /Form dictionary being a stream dictionary itself.

This is contrary to the spec, but Acrobat displays it without complaint.

Note: Acrobat pre-flight lists the resource to reference the form object
as "missing", and yet renders the form contents anyway.

Also note: we have a test file in the test suite that also includes "/Contents"
key, which is a metadata string, so we have to explicitly check that the value
for the "/Contents" key is a stream dictionary.