Support PDF files broken as per bug #702560
In this file a Form XObject dictionary is a simple dictionary, not a
stream dictionary as it is required to be. Instead the Form dictionary
has a /Contents key (as if this were a page) which references a stream
dictionary.

Acrobat, of course, opens the file without complaint.

So we do the same here, we note that a stream_offset of 0 means that
the Form dictionary is not a stream dictionary and we check to see if it
has a /Contents key whose value *is* a dictionary. If it is we merge
the two dictionaries and use the stream dictionary offset in the form
dictionary.

Add new error flag bits and message text.

We're going to have to start checking bug reports against pdfi and
adding any required extra handling.