graphics library - initialise some gstate members appropriately
The members are (I think) all part of the PDF graphics state rather than
the PostScript graphics state. These members are normally initialised
by the (PostScript) PDF interpreter but when we are running the new
pdfi interpreter inside Ghostscript we want to inherit the contents of
the graphics state, not overwrite them, so we do not initialise these.

This leads to problems; because the values are never initialised in this
case, and (for example) the text horizontal scaling remains at 0 when
it should be 100. Since we set the Text Rendering Matrix partially by
multiplying by that value, this can lead to a degenerate matrix and
cause text not to appear. In addition it causes ps2write output to be
incorrect and throw an error.

This commit does nothing with regular Ghostscript, but with GS+pdfi it
fixes hundreds of errors in ps2write tests and results in at least some
files drawing some content. At least one of these files is still not
correct (it is with pdfi alone) so there is clearly more to do here.

This reduces the number of errors testing GS+pdfi to 475 which is
447 errors less than Ghostscript alone.