graphics library - initialise more members of the graphics state
When creating a graphics state and when calling gs_initgraphics() the
code was not initialising several members of the graphics state. Mainly
these were the new members introduced some time back to allow the PDF
interpreter to track the PDF graphics state.

For consistency, and simpler implementation in pdfi, its better to
specifically initialise the state, especially those members which
should not default to 0.

This causes a few tiny diffs with XPS (because the XPS interpeter,
unlike the PS and PDF interpreters) did not set a couple of members
which are, nevertheless, used when rendering (smoothness and
accuratecurves). Smoothness (which affects gradient fills) is now
initialised to 0.02 instead of 1 (so much smoother) and accuratecurves
is now true instead of false.