Fix pdfwrite memory leak when doing multi-page output
When doing multiple page output (i.e. page-%0d.pdf or whatever) there
was a memory leak because it was attempting to short-circuit things and
just delete the last (blank) page, but there was a ton of stuff that
never got freed in that case.

This change sets a flag for the condition and then processes the whole
page (which hopefully doesn't do much because it is empty), including
freeing everything.  Then it just deletes the file at the end instead
of at the beginning of the function.

This function (pdf_close()) is over 1000 lines and really needs to be
refactored for maintainability.