Fix memory leak with pattern colour spaces
When setting a Pattern colour space we were using the graphics library
gs_setcolorspace() function instead of the pdfi-specific variant
pdfi_gs_setcolorspace().

The pdfi variant checks the current colour space and may take extra
action if the space being replaced is a Pattern space, in order to
free the pdfi objects.

If the current colour space was a Pattern, and we replaced it with a
Pattern space, this meant that we would leak the existing space.