Fix another pattern memory leak
In this case if we were in a pattern space, and had set a pattern, then
set a new pattern *without changing space*, then we would not clean
up the existing pattern. The previous commit fixed the case where we
were in a /Pattern space and then set a /Pattern space.