pdfwrite - fix memory leak with sclose
The stream code is hideous and full of surprised to trap the unwary. In
this case sclose() has the unexpected side-effect of setting the stream
buffer to NULL before returning, so we cannot free it afterwards.

s_close_filters deals with this for us by taking a pointer to the
buffer and calling gs_free_object on it after the stream is closed, so
we really need to do that (or mimic the same thing ourselves).

We could do s_close_filters(s, s->strm) to just close each filter in
turn, but I've chosen to just walk down the filter chain for each
filter we added, and then close them all up to that point.

This means changing pdf_end_encrypt() so that instead of closing the
stream (with the potential memory leak) we just signal whether the
calling code should close the encryption filter.