pdfwrite - review use of sclose().
The stream interface essentially leaves the buffer management in the
hands of the creator; sclose() does not free the stream buffer, but it
does set the pointer in the stream state to NULL.

This can be problematic; if the only reference we have to the original
buffer is the pointer in the stream state, then we must copy the pointer
before calling sclose() and then free the buffer afterwards.

s_close_filters() does this BUT it can't know whether a given buffer
was allocated in memory, from the C heap or some other fixed allocation.
It simply frees all the buffers. Obviously this can cause problems if
we use it indiscriminately.

I've reviewed all the places pdfwrite uses sclose() and where we can
use s_close_filters() I've modified the code to do so (to avoid memory
leaks in non-GC memory allocators). Where we must not attempt to free
the buffer I've left the sclose() but commented on the reason.