pdfwrite - free ICCBased profile streams to prevent memory leak
An oversight in memory cleanup in pdfwrite.

ICCBased colour space arrays contain a reference (must be a reference)
to a stream dictionary where the stream data is the profile. Because the
stream dictionary is a reference, it must have an ID (a pdf index). This
means that when we free the colour space array we do not free the stream
dictionary element contained within it.

We don't free objects with a non-zero ID because these are assumed to be
a type of resource, and tracked via the resource chains, but in the
case of ICC profiles we were not adding them to a resource chain.

Fixed here by adding ICC profile streams to the 'resourceOther' chain
so that they are freed on exit.