fix memory leak on error condition
When checking the contents of a Resources dictionary we expect each
value to also be a dictionary identified by type (XObject, Font etc).

The file wp12.pdf from bug #687396 has Resources dicitonaries where the
key is an empty name ('/') and the value is an indirect reference to an
illegal object (object number -1). This causes the dereferencing code
to return a null object.

The error checking treated an error return and an incorrect type as
equivalent, which meant we did not count down the key and value
retrieved from the dictionary which caused them to leak.

Use an alternative (existing) error label to count down the key and
value instead.