Bug 702587: Fix alpha buffer issue with fill/stroke operations.
The "alphabits" devices work by an alpha buffer being inserted
into the device chain before graphical operations. This intercepts
drawing device calls, 'saves' the color, and then converts the
call into a fill of a monochrome mask.

Only a small 'window' of mask is ever held in memory at once. When
the drawing moves out of the range of that window, the window is
flushed (conceptually the mask is scaled down, and then that scaled
representation is filled with the saved color).

Any "stragglers" (i.e. unwritten back bits of mask) are flushed at
the end of the graphical operation.

When using fill/stroke, the existing code fails to notice the
color in use being changed. This means the last 'window' of
fill can end up being written in the color intended for 'stroke'.

We therefore update the fill/stroke logic to check for the color
changing. (We check for 'unflushed data' and 'color has changed').