Fix bug in Type 4 function recursion depth
The depth was never getting decremented, so this would trigger an error on
functions that had lots of {} clauses, even if not nested.

The fix didn't show any visible difference on the test file, and no
cluster files seem to be affected, but this gets rid of the spurious
error.

Test file: bug 687401  pathsharingbug.pdf (not on the cluster)