Fix memory leaks with type 0 fonts
The error handling in pdfi_read_type_type0_font was incomplete,
leading to memory leaks. This commit reworks the error handling in the
function extensively, it now counts all the reference counted objects
down when exiting due to an error. If during the course of the function
we release one of these objects, we set its pointer to NULL. When
reference counted objects are stored in other structures we NULL the
pointer to transfer ownership to that structure.

Note that in a couple of cases we want to retain the pointer for a while
for convenient access to its members. These cases have been commented
regarding the ownership.