Add gs_param list parsing functions.
Actually, steal the existing functions from the PJL code,
move them into the graphics library, and refactor slightly.

2 functions; the first one, gs_param_list_add_tokens reads
a string of key/value pairs and adds them to the list.

The second one, gs_param_list_add_parsed_value takes a key
and a string for the value, and just adds that pair to the
list.