Fix memory leak with cidtype0 fonts
The type 0 font code calls either pdfi_read_cidtype2_font or
pdfi_read_cidtype0_font and in both cases hands in a memory buffer which
it created.

The type 2 case stores the pointer in the descpfont structure which it
returns and manages the buffer lifetime as the lifetime of that object.

The type 0 case (currently) just returns an error. Since nothing took
ownership of the buffer it leaked.

In addition it was possible for the type 2 case to return an error
before storing the pointer in which case it would also leak.

The ownership is now explicit, at the point of calling the 'read'
functions the buffer belongs to, and is managed by, the called 'read'
function.