Fix memory leak in patterns with shading and functions
Really obscure sort of leak, but there was a pattern that uses a
shading and a separation colorspace that has a function.  The function from
the separation colorspace was leaking because when it got freed it was
down in the gstate code freeing the pattern's saved gstate, so none of our
hacky code that ensured the colorspaces got cleaned up would get called.

Added pdfi_colorspace_cleanup() so we can do the cleanup on a
specific colorspace.  Call this in pdfi_pattern_cleanup() on the
colorspaces in the 'saved' gstate in case those are about to be freed.

Also for good measure, also call pdfi_colorspace_cleanup() in
pdfi_shading_free() before decrementing the colorspace there.  But
that wasn't the issue on the sample I was debugging.

Sample is: Bug 687401 pathsharingbug.pdf (not on cluster)
But I think this leak was probably in lots of files.