fix memory leaks with encrypted files
When adding the decryption and SubFileDecode filters we must ensure that
we set the 'original' member of the pdf_stream object properly so that
when we close the pdf_stream we properly close all the filters which
we added.

In addition, we can't pdfi_countdown() pdf_stream objects because these
are not reference counted objects! We need to close the file and then
call gs_free_object to free the memory.