Fix Bug 702598: Ghostscript will not open SECURED PDF that Adobe opens.
It's not really clear that this is a valid PDF since it does not conform to
one comment in the spec related to AuthEvent with StmF and StrF /Identity,
but since Adobe opens this we now open PDF's with AuthEvent specified as
something other than /DocOpen and BOTH StrF and StmF /Identity. We look
in the /CF "StdCF" dictionary for the AuthEvent. As per the spec., if
AuthEvent is not present, assume DocOpen, and if StmF or StrF are not
present, we default to /Identity.

Thanks to Ken Sharp for the initial work on this.