Fix memory leaks with Indexed colour spaces
When freeing the pdfi interpreter portions of a colour space prior to
the graphics library freeing it we were not checking to see if the
current colour space was /Indexed. If it is then we need to look at the
base space, in case that is a space which we need to clean up.