Memory leak fixes with gs_setgstate()
The gs_setgstate function replaces the contents of the current graphics
state, but because it does not know about our extra interpreter
storage, it does not reference count the fonts correctly.

To cater for this add a new pdfi_gs_setgstate to be used instead; this
*must* be used in combination with a pdfi_gsave before calling, and a
pdfi_grestore after. It counts down the PDF_font object in the current
graphics state (if there is a current font) before calling gs_setgstate
and counts up the PDF_font object after gs_setgstate (if present).

This keeps the font reference counts straight and prevents memory leaks
or crashes.