Added thread safety to the viewer.
The viewer now ensures that no more than one ghostscript call is
happening at the same time and has multiple options for controlling
the behavior if multiple operations are attempting to be called.
Also began work on a utility class to make Ghostscript calls easier
by having a class carry around a instance and a caller handle.