Fix some loop detection problems in type 5 halftones
We were erroneously detecting circular references because we were
not putting a mark and clear round the initial extraction of an halftone
from the dictionary. If any of the subsequent inks used the same
actual halftone dictionary as the initial ink, we would detect this as
a circular reference.

This fixes the warning, some memory leaks and causes a progression in
the cluster test file CATX1181.pdf with the psdcmyk device.