Fix memory leak with shading, smask
(includes adding gstate parameter to client callback, from master
249da10b83dd8ed287877e6d21fa3ca3658aadea)

Now pdfi does the color cleanup in pdfi_gstate_free_cb() which is
while the gstate is being freed (instead of before).

This fixes issues where a pattern wasn't getting cleaned up because
it was decremented down in pdfi_gstate_smask_free() and the refcnt
was 2 at that time (and was also 2 in pdfi_grestore()).

Also changed the pdfi_color_cleanup() api to pass in the gstate
explicitly instead of assuming it's the one from the context.

Sample: Bug 690208