Fix memory leaks with CMaps
It seems Acrobat will accept technically invalid CMaps which are missing
most of the PostScript boilerplate (comments, CIDInit ProcSet) and in
particular do not have a CMapType entry.

The lack of a CMapType left the pdf_cmap object with a type of 0, and
the cleanup code only processed the contents if the type was 1, thus
leading to the content of the CMap and the CMap itself leaking.

For now (because we do not yet process ToUnicode CMaps) simply set the
type to 1 before parsing the CMap stream.