Fix array element type checking
When checking the type of an array in an element we must retrieve the
object, but if its not the right type, do not update the object
pointer.

This was leading to memory corruption because we were counting down
objects too early and freeing them when there was an error.