Fix Bug 702472: Array contents scrambled with writeobject/printobject
Thanks to Peter Cherepanov for this patch.

The mix-up of the nested arrays was caused by depth-first traversal of
the array tree and assignment of offsets to subsidiary arrays that assumed
width-first traversal. Switching to width-first traversal in all cases is
the easiest way to fix the problem.

Additionally, access to the list of arrays is now implemented with the
-index- operator that has constant complexity, rather than with -roll-
that has linear complexity. So the whole process now has linear complexity
rather than quadratic.

The total number of arrays in the exported object is arbitrarily limited to
64K. This prevents memory exhaustion when exported object has circular
references, but should not hinder real world usage, if any.

Following the original implementation, equal names are not merged and
exported separately.

NOTE:
A Ghostscript extension, support of dictionaries in binary object sequence
has been removed. Following the recent trend of cleaning up non-standard
features.

Dictionary support is not yet removed from the C part. Contrary to PLRM,
Adobe can parse binary object sequences with a non-zero tag in the top
object. This feature is not yet added. Sequences that contain zero
length names can be exported but cannot be imported due a conflict with
accessing the user name index. This is a design flaw in PostScript.