Improve client memory handling for colour spaces and patterns
Colour spaces have had a client_data member for some time (only used
with Pantone colour replacement) but no means to signal to the client
that a colour space was going to be freed.

In addition, its terribly hard for an interpreter which uses complex
colour spaces (such as Separation and DeviceN, for Patterns see below)
to reliably track when a colour space is freed.

This commit adds a new 'interpreter_data' pointer to hold data allocated
by the interpreter which is associated with a colour space. We also add
an 'interpreter_free_cspace_proc' to be called when a colour space is
about to be freed. This gives the interpreter an opportunity to clean
up the memory it has associated with the colour space.

Similarly pattern instances (as opposed to Pattern colour spaces) have
long had a client_data member, which was previously stored in the
'template' (which is a terrible misnomer) for both type 1 and type 2
patterns. However there was no procedure to notify the client when the
pattern was freed! The PCL interpreter (which was the only one using
this type of pattern) dealt with this by replacing the 'free' routine
in the structure with its own routine which first freed the client data
and then called the original free routine. This seems like a hack.

This commit moves the client_data out of the pattern instance and into
the pattern structure proper. It also adds a 'notfy_free' member which
holds a pointer to a function to be called when a pattern is freed. This
gives the client a reliable opportunity to clean up the client_data.

These changes involve updating the PostScript, PCL and XPS interpreters
to point to the new location of the client_data, as well as similar
changes in the graphics library. pixmap patterns now use the notify_free
functio0n instead of patching over the structure's free routine.

Update pdfi to use the new colour space and Pattern free notifications

Now that the graphics library will tell us when colour spaces and
pattern instances are about to be freed, have the pdfi interpreter
supply those functions and use them.

This absolves us of an awful lot of tedious, hairy and frankly less
than reliable book-keeping trying to keep track of colour space and
patterns through gsave/grestore etc.

This commit should be squashed up with the preceding one and committed
as one. They were separated in order to do simplify committing the
non-pdfi changes to the master branch.

There may well be more simplification possibilities, these will be
undertaken later.

Also squashed into this commit; fix reference counting problem spotted
by Nancy with the original commit.